services/featureflags/pyproject.toml
[project]
name = "featureflags"
version = "0.1.0"
description = "Feature Flags Service"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.0",
    "dj-database-url",
    "gunicorn",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app", "src/featureflags", "src/core"]
